refactor: decompose tokenList components
