+++ apps/cowswap-frontend/src/modules/tokensList/pure/SelectTokenModal/index.tsx
@@ -74,6 +74,7 @@ export function SelectTokenModal(props: SelectTokenModalProps): ReactNode {
     unsupportedTokens,
     permitCompatibleTokens,
     tokenListTags,
+    isWalletConnected: !!account,
   }
 
   const trimmedInputValue = inputValue.trim()
